refactor: GameScene 리팩토링
- 코드의 가독성을 위해 GameScene 한파일에 모두 정의 되었던 Node들을 GameScene+Extension 파일을 새로 생성하여 Node들을 정의 하는 코드들을 정리 하였습니다.
- 주석의 내용을 수정하고 필요한 경우 가독성을 위해 구분선을 추가하였습니다.
